% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launder.R
\name{common_features}
\alias{common_features}
\title{Common features in raw data}
\usage{
common_features(features)
}
\arguments{
\item{features}{list of matrices.}
}
\value{
A three-dimensional array containing the common features
(as per column names).
}
\description{
Extracts common features from heterogenous data and simplifies them into a
three-dimensional array, thus making them homogenous.
}
\seealso{
Other data laundering techniques: 
\code{\link{cut_off}()},
\code{\link{graphed_features}()},
\code{\link{induced.subgraph.by_name}()},
\code{\link{launder}()}
}
\concept{data laundering techniques}
