% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launder.R
\name{cut_off}
\alias{cut_off}
\title{Remove lightweight edges}
\usage{
cut_off(graph, threshold = NaN, threshold.quantile = NaN)
}
\arguments{
\item{graph}{The original graph}

\item{threshold}{numeric. The minimal weight of an edge to keep.}

\item{threshold.quantile}{numeric. The minimal relative weight of an edge
to keep with respect to all other edge weights.}
}
\value{
A subgraph of \code{graph}, where edges have been removed according
to the given threshold. If no threshold is given \code{graph} is returned
unchanged.
}
\description{
Cuts off edges below a threshold.
}
\seealso{
Other data laundering techniques: 
\code{\link{common_features}()},
\code{\link{graphed_features}()},
\code{\link{induced.subgraph.by_name}()},
\code{\link{launder}()}
}
\concept{data laundering techniques}
