% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{edge_importance}
\alias{edge_importance}
\title{Calculate the importance of particular edges in the graph.}
\usage{
edge_importance(graph, trees, tree_importances, mc.cores = 1, sep = "$")
}
\arguments{
\item{graph}{The graph to analyse}

\item{trees}{The decision trees returned by DFNET iterations}

\item{tree_importances}{Numeric importance scores assigned to each tree}

\item{mc.cores}{how many cores to use in parallel}

\item{sep}{The seperator used to flatten the features from which
\code{trees} were generated.}
}
\value{
the importance of each edge in \code{graph} w.r.t. \code{trees}.
}
\description{
Calculate the importance of particular edges in the graph.
}
\seealso{
Other metrics: 
\code{\link{feature_importance}()},
\code{\link{module_importance}()},
\code{\link{unique_module_importance}()}
}
\concept{metrics}
