% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{feature_importance}
\alias{feature_importance}
\title{Importance of individual features}
\usage{
feature_importance(forest, features, sep = "$")
}
\arguments{
\item{forest}{The trained \code{DFNET.forest}.}

\item{features}{matrix or 3D array. The features on which it was trained.}

\item{sep}{string.}
}
\value{
A matrix of importance scores with one row per column in
\code{features} and one column per matrix (in a 3D array).
}
\description{
Calculates the average importance of features over multiple decision
trees.
}
\seealso{
Other metrics: 
\code{\link{edge_importance}()},
\code{\link{module_importance}()},
\code{\link{unique_module_importance}()}
}
\concept{metrics}
