% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten2ranger}
\alias{flatten2ranger}
\title{Flatten multi-modal data for ranger calls.}
\usage{
flatten2ranger(data, cols, sep = "$")
}
\arguments{
\item{data}{matrix or 3D array. The data to flatten.}

\item{cols}{vector. The (optional) columns to use.}

\item{sep}{string. A separator to use when building the new column names.}
}
\value{
A matrix containing \code{data} reduced to \code{cols}, with
the third dimension inserted as extra columns.
}
\description{
Produces a flat matrix from \code{data}, a potentially 3-dimensional array.
The column names of the new matrix are produced by pasting the dimnames
together, using \code{sep} as separator.
}
\keyword{internal}
