% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launder.R
\name{graphed_features}
\alias{graphed_features}
\title{Retain meaningful features}
\usage{
graphed_features(features, graph)
}
\arguments{
\item{features}{matrix or 3D array.}

\item{graph}{an igraph.}
}
\value{
A matrix or 3D array (as features \code{features}), reduced and
reshaped such that only features belonging to the vertices in \code{graph}
appear in the order given by \code{graph}.
}
\description{
Trims \code{features} so that only columns whose names match the vertex names
in \code{graph} are retained.
}
\seealso{
Other data laundering techniques: 
\code{\link{common_features}()},
\code{\link{cut_off}()},
\code{\link{induced.subgraph.by_name}()},
\code{\link{launder}()}
}
\concept{data laundering techniques}
