% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{head.DFNET.forest}
\alias{head.DFNET.forest}
\title{Return the first trees in the forest}
\usage{
\method{head}{DFNET.forest}(x, n.generations = 6L, ...)
}
\arguments{
\item{x}{The original \code{DFNET.forest}.}

\item{n.generations}{integer. The number of generations to keep.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
The first \code{n.generations} generations worth of modules
and trees.
}
\description{
Returns the first (= oldest) trees in the forest.  The number of trees to
keep is determined by multiplying the number of generations with the
generation size of the forest (the \code{ntrees} parameter)
in \link{train}.
}
\details{
This function should not be used to retrain the forest from an earlier
branch.  It does not return \code{last.performance} and may return a wrong
value for \code{walk.depth}.
}
