% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launder.R
\name{induced.subgraph.by_name}
\alias{induced.subgraph.by_name}
\title{Subgraph of a graph}
\usage{
induced.subgraph.by_name(graph, names)
}
\arguments{
\item{graph}{The original graph.}

\item{names}{vector of strings. The names of the vertices which will form
the subgraph.}
}
\value{
The induced subraph.
}
\description{
Creates a subgraph of a graph, containing only the specified vertices and
the edges between them.
}
\seealso{
\link[igraph:induced.subgraph]{igraph::induced_subgraph()}, which
this procedure uses internally.

Other data laundering techniques: 
\code{\link{common_features}()},
\code{\link{cut_off}()},
\code{\link{graphed_features}()},
\code{\link{launder}()}
}
\concept{data laundering techniques}
\keyword{internal}
