% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{init}
\alias{init}
\title{Model initialization}
\usage{
init(
  graph,
  features,
  target,
  ntrees = 100,
  walk.depth = NaN,
  performance = NULL,
  flatten.sep = "$",
  importance = "impurity",
  splitrule = "gini"
)
}
\arguments{
\item{graph}{The graph to train the network on.}

\item{features}{numeric matrix or 3D array. The features to train on.}

\item{target}{numeric vector. The target to train towards.}

\item{ntrees}{integer. The number of trees to generate per iteration.}

\item{walk.depth}{integer. The number of nodes to select per module.}

\item{performance}{unary function. Called with a decision tree as argument to
estimate that tree's performance.}

\item{flatten.sep}{string. Separator to use when flattening features.}

\item{importance}{variable importance mode.
See \link{ranger:ranger}{ranger::ranger}.}

\item{splitrule}{Splitting rule.
See \link{ranger:ranger}{ranger::ranger}.}
}
\value{
An initialized \code{DFNET.forest}.
}
\description{
Initializes the decision forest network.
}
\keyword{internal}
