% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launder.R
\name{launder}
\alias{launder}
\title{Data laundering}
\usage{
launder(graph, features, threshold = NaN, threshold.quantile = NaN)
}
\arguments{
\item{graph}{The original graph.}

\item{features}{list of matrices.}

\item{threshold}{see \link{cut_off}.}

\item{threshold.quantile}{see \link{cut_off}.}
}
\value{
A list \code{graph, features} with the laundered graph and features.
}
\description{
Launders input graph and features.
}
\seealso{
Other data laundering techniques: 
\code{\link{common_features}()},
\code{\link{cut_off}()},
\code{\link{graphed_features}()},
\code{\link{induced.subgraph.by_name}()}
}
\concept{data laundering techniques}
