% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{module_importance}
\alias{module_importance}
\title{Calculate the importance of particular modules in the graph.}
\usage{
module_importance(
  graph,
  modules,
  edge_importances,
  tree_importances,
  mc.cores = 1
)
}
\arguments{
\item{graph}{the graph used for training}

\item{modules}{the modules obtained by training}

\item{edge_importances}{the importance scores assigned to each edge in the
graph, e.g. by \code{edge_importance}.}

\item{tree_importances}{the importance scores assigned to each decision tree
in the forest}

\item{mc.cores}{how many cores to use in parallel}
}
\value{
the accumulated edge importance and total importance of each module.
}
\description{
Calculate the importance of particular modules in the graph.
}
\seealso{
Other metrics: 
\code{\link{edge_importance}()},
\code{\link{feature_importance}()},
\code{\link{unique_module_importance}()}
}
\concept{metrics}
