% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.DFNET.forest}
\alias{predict.DFNET.forest}
\title{Model predictions}
\usage{
\method{predict}{DFNET.forest}(object, data, ...)
}
\arguments{
\item{object}{The \code{DFNET.forest} to use for prediction.}

\item{data}{matrix or 3D array. The data to run predictions on.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list with members
\describe{
  \item{predictions}{the predicted classes}
  \item{approval.rate}{the relative number of decision trees that predicted
the given class.}
  \item{participation.rate}{the relative number of decision trees that voted
at all (i.e. did not predict \code{NaN} or \code{NA})}.
}
}
\description{
Uses a \code{DFNET.forest} to run predictions on data.
}
\examples{
\dontrun{
smp_size <- floor(0.80 * dim(features)[1])
train_ids <- sample(dim(features)[1], size = smp_size)
test_ids <- (1:dim(features)[1])[-train_ids]
forest <- DFNET_init(graph, features[train_ids, ])
# train ...
prediction <- predict(forest, features[test_ids, ])
}
}
