% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launder.R
\name{relat}
\alias{relat}
\title{Relativize data}
\usage{
relat(x, na.rm = FALSE, default = 1, default.na = NaN)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{logical. Should NaNs be removed from \code{x} when calculating
minimal and maximal values?}

\item{default}{numeric. The default value(s) to use when all values in
\code{x} are equal. Should lie between 0 and 1.}

\item{default.na}{numeric. As default, but applied if minimal or maximal
value are NaN.  Not applicable if \code{na.rm} is true.}
}
\value{
The relativized vector \eqn{\alpha}, such that
\eqn{x = (1-\alpha)\min(x) + \alpha\max(x)} holds w.r.t. floating point
inaccuracies.
}
\description{
Maps the elements of a numeric vector to the interval [0:1].
}
\examples{
relat(1:3)
relat(c(NaN, 2, 3, 4))
relat(c(NaN, 2, 3, 4), na.rm = TRUE)
}
