% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tail.DFNET.forest}
\alias{tail.DFNET.forest}
\title{Return the last trees in the forest}
\usage{
\method{tail}{DFNET.forest}(x, n.generations = 6L, ...)
}
\arguments{
\item{x}{The original \code{DFNET.forest}.}

\item{n.generations}{integer. The number of generations to keep.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
The last \code{n.generations} generations worth of modules
and trees.
}
\description{
Returns the last (= newest) trees in the forest.  The number of trees to
keep is determined by multiplying the number of generations with the
generation size of the forest (the \code{ntrees} parameter)
in \link{train}.
}
\details{
This function can be used to shrink the forest while training (since only
the last generation will be used anyway).
}
