% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tester}
\alias{tester}
\title{Higher-order model testing}
\usage{
tester(data, target, metric = ModelMetrics::auc, do.predict = TRUE)
}
\arguments{
\item{data}{matrix or 3D array. The data to use for testing.}

\item{target}{numeric vector. The target values to predict.}

\item{metric}{binary function. A metric to compare actual values
with predictions.}

\item{do.predict}{logical. Should \code{predict} be called (TRUE) or
\code{$predictions} be accessed directly (FALSE)?}
}
\value{
A unary function which, takes a ranger-like predictor and evaluates
\code{metric} with \code{target} and the predictions of \code{predictor}
on \code{data}.
}
\description{
Returns a procedure that can be used as \code{performance} metric in
\link{train}.
}
