% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{unique_module_importance}
\alias{unique_module_importance}
\title{Calculate the importance of each unique module.}
\usage{
unique_module_importance(
  graph,
  modules,
  edge_importances,
  tree_importances,
  mc.cores = 1,
  collapse = mean
)
}
\arguments{
\item{graph}{the graph used for training}

\item{modules}{the modules obtained by training}

\item{edge_importances}{the importance scores assigned to each edge in the
graph, e.g. by \code{edge_importance}.}

\item{tree_importances}{the importance scores assigned to each decision tree
in the forest}

\item{mc.cores}{how many cores to use in parallel}

\item{collapse}{how accumulate multiple values per module}
}
\value{
the collapsed tree and edge importances of each unique module, as
well as their sum.
}
\description{
Calculate the importance of each unique module.
}
\seealso{
Other metrics: 
\code{\link{edge_importance}()},
\code{\link{feature_importance}()},
\code{\link{module_importance}()}
}
\concept{metrics}
